backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow
keras
opencv-python
python-jose
passlib
python-dotenv
pydantic
requests
